split util, use util in message
